+++ app/src/main/java/org/example/models/RestaurantReviewModel.java
@@ -3,7 +3,7 @@ package org.example.models;
 public class RestaurantReviewModel extends Review {
   private RestaurantModel restaurant;
 
-  public RestaurantReviewModel(String reviewerName, Integer rating, String comment, RestaurantModel restaurant) {
+  public RestaurantReviewModel(String reviewerName, Double rating, String comment, RestaurantModel restaurant) {
     super(reviewerName, rating, comment);
     this.restaurant = restaurant;
   }
